+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/response/WalkGetDetailResponseDto.java
@@ -39,16 +39,16 @@ public class WalkGetDetailResponseDto {
     String themeName;
 
     @Schema(description = "출발지 위도")
-    BigDecimal startLatitude;
+    double startLatitude;
 
     @Schema(description = "출발지 경도")
-    BigDecimal startLongitude;
+    double startLongitude;
 
     @Schema(description = "도착지 위도")
-    BigDecimal endLatitude;
+    double endLatitude;
 
     @Schema(description = "도착지 경도")
-    BigDecimal endLongitude;
+    double endLongitude;
 
     @Schema(description = "산책 경로")
     List<RouteGetResponseDto> routeList;
